pyproject.toml
[tool.poetry]
name = "datenspende-who5"
version = "0.1.0"
description = ""
authors = ["Marc Wiedermann <marcwiedermann@posteo.de>"]
readme = "README.md"
packages = [{include = "datenspende_who5"}]

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
pandas = "^2.0.0"
pyarrow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
